Update install.sh
& version numbers across all files.

+++ install.sh
@@ -1,16 +1,29 @@
-#!/bin/sh
+ #!/bin/bash
 
-echo "Installer script for sys2mqtt v0.2.9"
-echo ""
-echo "Copying and enabling sys2mqtt service"
+ echo "sys2mqtt v0.3.0"
+ echo "sys2mqtt will now be installed as a systemd service, if you do not use systemd this script will not work"
+ echo "Would you like to continue? (y/n)"
+ read var_cont
 
-cp sys2mqtt.service  /etc/systemd/system/sys2mqtt.service
-
-systemctl daemon-reload
-systemctl enable sys2mqtt.service
-systemctl start sys2mqtt.service
-
-echo ""
-echo "Service has now been enabled and started"
-echo "To change configuration, issue 'systemctl stop sys2mqtt' before editing"
-echo "the conf.py file."
+ if [ $var_cont = n ]
+ then
+    echo "You have chosen not to continue."
+    echo "Visit the sys2mqtt wiki for alternative installation methods"
+    exit 1
+    
+ elif [ $var_cont = y ]
+ then
+    echo "sys2mqtt is now installing..."
+    echo "Installing dependencies..."
+    apt install python3 python3-pip -y
+    echo "Installing and enabling sys2mqtt service"
+    cp sys2mqtt.service /etc/systemd/system/sys2mqtt.service
+    systemctl daemon-reload
+    systemctl enable sys2mqtt.service
+    echo "Starting sys2mqtt"
+    systemctl start sys2mqtt.service
+    echo "sys2mqtt has been installed and started.  It will automatically start after every reboot."
+    exit 0
+else
+    echo "That isn't a valid choice, try starting the script again."
+    exit 2
